+++ packages/config/src/config-worker.ts
@@ -1,13 +1,12 @@
-import { loadConfig } from 'unconfig'
 import type { Options } from './options'
 
 export async function loadConfigAsync(cwd: string): Promise<Options> {
+  const { loadConfig } = await import('unconfig')
   const { config } = await loadConfig<Options>({
     sources: [
       {
         files: 'vue-macros.config',
         extensions: ['mts', 'cts', 'ts', 'mjs', 'cjs', 'js', 'json', ''],
-        loader: 'jiti',
       },
       {
         files: 'package.json',