+++ packages/script-lang/tests/fixtures.test.ts
@@ -1,9 +1,9 @@
 import { resolve } from 'node:path'
 import {
   rollupBuild,
-  RollupEsbuildPlugin,
   RollupVue,
   testFixtures,
+  UnpluginOxc,
 } from '@vue-macros/test-utils'
 import { describe } from 'vitest'
 import VueScriptLang from '../src/rollup'
@@ -18,9 +18,7 @@ describe('fixtures', async () => {
           isProduction: args.isProduction,
           inlineTemplate: args.isProduction,
         }),
-        RollupEsbuildPlugin({
-          target: 'esnext',
-        }),
+        UnpluginOxc.rollup(),
       ]),
     {
       cwd: resolve(__dirname, '..'),