+++ packages/setup-component/tests/fixtures.test.ts
@@ -1,10 +1,10 @@
 import { resolve } from 'node:path'
 import {
   rollupBuild,
-  RollupEsbuildPlugin,
   RollupVue,
   RollupVueJsx,
   testFixtures,
+  UnpluginOxc,
 } from '@vue-macros/test-utils'
 import { describe } from 'vitest'
 import VueSetupComponent from '../src/rollup'
@@ -17,9 +17,7 @@ describe('fixtures', async () => {
         VueSetupComponent(),
         RollupVue(),
         RollupVueJsx(),
-        RollupEsbuildPlugin({
-          target: 'esnext',
-        }),
+        UnpluginOxc.rollup(),
       ]),
     {
       cwd: resolve(__dirname, '..'),