+++ packages/short-bind/tests/fixtures.test.ts
@@ -1,9 +1,9 @@
 import { resolve } from 'node:path'
 import {
   rollupBuild,
-  RollupEsbuildPlugin,
   RollupVue,
   testFixtures,
+  UnpluginOxc,
 } from '@vue-macros/test-utils'
 import { describe } from 'vitest'
 import ShortBind from '../src/rollup'
@@ -12,13 +12,7 @@ describe('fixtures', async () => {
   await testFixtures(
     'tests/fixtures/*.{vue,[jt]s?(x)}',
     (args, id) =>
-      rollupBuild(id, [
-        ShortBind(),
-        RollupVue(),
-        RollupEsbuildPlugin({
-          target: 'esnext',
-        }),
-      ]),
+      rollupBuild(id, [ShortBind(), RollupVue(), UnpluginOxc.rollup()]),
     {
       cwd: resolve(__dirname, '..'),
       promise: true,